Separate cache structs into their own files
